+++ HeritageEntityFramework/Item.cs
@@ -16,6 +16,6 @@ namespace HeritageEntityFramework
     {
         public int Id { get; set; }
         public string Description { get; set; }
-        public long Price { get; set; }
+        public int Price { get; set; }
     }
 }
